W-12049491: Added new library to support Draft 6, 7, 2019-09 and 2020-12
