+++ src/main/java/org/mule/module/json/internal/JsonSchemaParser.java
@@ -1,12 +1,5 @@
 package org.mule.module.json.internal;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.mule.runtime.extension.api.exception.ModuleException;
-
-import java.net.URL;
-
 import static com.google.common.base.Preconditions.checkState;
 import static org.mule.module.json.api.JsonError.INVALID_INPUT_JSON;
 import static org.mule.module.json.api.JsonError.SCHEMA_NOT_FOUND;
@@ -14,9 +7,15 @@ import static org.mule.module.json.internal.ValidatorCommonUtils.isBlank;
 import static org.mule.module.json.internal.ValidatorCommonUtils.resolveLocationIfNecessary;
 import static java.lang.String.format;
 
+import org.mule.runtime.extension.api.exception.ModuleException;
+import java.net.URL;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 public class JsonSchemaParser {
 
-    ObjectMapper objectMapper;
+    private final ObjectMapper objectMapper;
 
     public JsonSchemaParser(){
         objectMapper = new ObjectMapper();
@@ -26,7 +25,7 @@ public class JsonSchemaParser {
      * Load Json node from schema content or schema location, and make validations.
      * Decided to obtain first the JsonNode, to read the json and know what version of
      * schema is needed and then create JsonSchemaValidator with class JsonSchema of
-     * com.github.fge (Draft 3 & 4) or com.networknt (Draft 6, 7, 2019-09 & 2020-12)
+     * java-json-tools (Draft 3 & 4) or com.networknt (Draft 6, 7, 2019-09 & 2020-12)
      */
 
     public JsonNode getSchemaJsonNode(String schemaContent, String schemaLocation) {
@@ -42,12 +41,8 @@ public class JsonSchemaParser {
             checkState(schemaLocation != null, "schemaLocation has not been provided");
             return objectMapper.readTree(new URL(resolveLocationIfNecessary(schemaLocation)));
         } catch (Exception e) {
-
             throw new ModuleException(format("Could not load JSON schema [%s]. %s", schemaLocation, e.getMessage()),
                     SCHEMA_NOT_FOUND, e);
         }
     }
-
-
-
 }
