+++ src/main/java/org/mule/module/json/internal/ValidateJsonSchemaOperation.java
@@ -6,18 +6,14 @@
  */
 package org.mule.module.json.internal;
 
-import static com.networknt.schema.SpecVersion.VersionFlag.*;
-import static java.lang.String.format;
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.stream.Collectors.toMap;
 import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
 import static org.mule.runtime.api.meta.ExpressionSupport.NOT_SUPPORTED;
 import static org.mule.runtime.api.meta.model.operation.ExecutionType.CPU_INTENSIVE;
 import static org.mule.runtime.extension.api.annotation.param.display.Placement.ADVANCED_TAB;
+import static java.lang.String.format;
+import static java.util.concurrent.TimeUnit.MINUTES;
+import static java.util.stream.Collectors.toMap;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.networknt.schema.JsonSchemaException;
-import com.networknt.schema.SpecVersionDetector;
 import org.mule.module.json.api.JsonSchemaDereferencingMode;
 import org.mule.module.json.api.SchemaRedirect;
 import org.mule.module.json.internal.cleanup.JsonModuleResourceReleaser;
@@ -64,6 +60,7 @@ import org.apache.commons.pool2.impl.GenericObjectPoolConfig;
 import org.mule.runtime.extension.api.exception.ModuleException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import com.fasterxml.jackson.databind.JsonNode;
 
 /**
  * Operation to validate an XML document against a schema
@@ -76,8 +73,7 @@ public class ValidateJsonSchemaOperation implements Disposable, Startable, Stopp
   private static final int MAX_IDLE_POOL_COUNT = 32;
   private final static Logger LOGGER = LoggerFactory.getLogger(ValidateJsonSchemaOperation.class);
   private JsonModuleResourceReleaser resourceReleaser;
-  private JsonSchemaParser jsonSchemaParser = new JsonSchemaParser();
-  private ValidatorSchemaLibraryDetector validatorSchemaLibraryDetector = new ValidatorSchemaLibraryDetector();
+  private final JsonSchemaParser jsonSchemaParser = new JsonSchemaParser();
 
   @Inject
   TransformationService transformationService;
@@ -225,29 +221,14 @@ public class ValidateJsonSchemaOperation implements Disposable, Startable, Stopp
     return new BasePooledObjectFactory<JsonSchemaValidator>() {
 
       @Override
-      public JsonSchemaValidator create() throws Exception {
+      public JsonSchemaValidator create() {
 
-        // Obtengo el schema como jsonNode
         JsonNode schemaJsonNode = jsonSchemaParser.getSchemaJsonNode(key.schemaContent, key.schemas);
 
-        org.mule.module.json.internal.Validator validator;
-
-        if (validatorSchemaLibraryDetector.detectValidator(schemaJsonNode).equals(ValidationLibraries.NETWORKNT)) {
-          validator =
-                  new ValidatorNetworknt().create(schemaJsonNode, key.schemas, key.dereferencingType, key.schemaRedirects, key.schemaContent);
-        } else {
-          validator = new ValidatorJavaJsonTools().create(schemaJsonNode, key.schemas, key.dereferencingType, key.schemaRedirects, key.schemaContent);
+        if (ValidatorSchemaLibraryDetector.detectValidator(schemaJsonNode).equals(ValidationLibraries.NETWORKNT)) {
+          return new SchemaValidatorNetworknt(key.schemas, key.dereferencingType, key.allowDuplicateKeys, key.allowArbitraryPrecision, key.schemaRedirects, schemaJsonNode);
         }
-
-        return JsonSchemaValidator.builder()
-                .setValidator(validator)
-            .addSchemaRedirects(key.schemaRedirects)
-            .setDereferencing(key.dereferencingType)
-            .setSchemaLocation(key.schemas)
-            .allowDuplicateKeys(key.allowDuplicateKeys)
-            .allowArbitraryPrecision(key.allowArbitraryPrecision)
-            .setSchemaContent(key.schemaContent)
-            .build();
+        return new SchemaValidatorJavaJsonTools(key.schemas, key.dereferencingType, key.allowDuplicateKeys, key.allowArbitraryPrecision, key.schemaRedirects, schemaJsonNode);
       }
 
       @Override
