--- src/main/java/org/mule/module/json/internal/pepe.java
@@ -1,4 +1,0 @@
-package org.mule.module.json.internal;
-
-public class pepe {
-}
